+++ changelog.md
@@ -1,1 +1,2 @@
 - Fixed crash on newer NeoForge versions
+- Re-added player rendering to the obituary screen